backend/app/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
groq         # if you use Groq SDK
websockets   # if you use websockets
